+++ src/Components/Pages/Visualizer/Graphs/MVT/index.js
@@ -5,11 +5,11 @@ import { anglesRightIcon } from "../../../../../Resources/Icons";
 import { MvTSelectButton } from "../../../../Shared/Buttons";
 import { Icon } from "../../../../Shared/Icon";
 import { FontVarTitle } from "../../../../Shared/Title";
-import ResizableContainer from "../GraphContainer";
+import ResizableContainer from "../Components/GraphContainer";
 import Manipulator from "./Manipulator";
 import MvtGraph from "./Graph";
 import { mvtGraphComputation } from "../../Computation/MVT";
-import { dummyData } from "../../Data/data";
+import { dummyData } from "../../Ancilliary/Data/data";
 import { VizDataHistoryContext } from "../../../../../Context/visualizer";
 
 const LABEL_TOGGLES = { "Replica 1": true, "Replica 2": true, "Replica 3": true, "Replica 4": true }
