+++ src/Components/Pages/Visualizer/Table/Components/SmallTable.js
