+++ src/Components/Pages/Visualizer/Table/index.js
@@ -2,9 +2,9 @@ import classNames from 'classnames';
 import React, { Fragment, useContext, useEffect, useState } from 'react';
 import { DATA_TABLE_DELAY, DATA_TABLE_NO_PRIMARY_EXISTS } from '../../../../Constants';
 import { computeTableData } from '../Computation/TransInfo';
-import { tableDataDummy } from '../Data/data';
+import { tableDataDummy } from '../Ancilliary/Data/data';
 import { FontVarTitle } from '../../../Shared/Title';
-import Carousel from './Carousel';
+import Carousel from './Components/Carousel';
 import { VizDataHistoryContext } from '../../../../Context/visualizer';
 
 const TABLE_HEADERS = {
@@ -202,8 +202,8 @@ const DataTable = ({ delay = DATA_TABLE_DELAY }) => {
                         </tbody>
                     </table>
                 </div>
-                <Carousel 
-                    onPrev={() => handleChangePage('prev')} onNext={() => handleChangePage('next')} nextDisabled={currentPage === totalPages} prevDisabled={currentPage === 1} 
+                <Carousel
+                    onPrev={() => handleChangePage('prev')} onNext={() => handleChangePage('next')} nextDisabled={currentPage === totalPages} prevDisabled={currentPage === 1}
                     startRecord={startRecord}
                     endRecord={endRecord}
                     data={tableData}
